# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/reportes/psl": {
      "get": {
        "operationId": "ReportesController_getPslReport",
        "parameters": [
          {
            "name": "desde",
            "required": true,
            "in": "query",
            "description": "YYYY-MM-DD Colombia local",
            "schema": {
              "example": "2026-07-01",
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "required": true,
            "in": "query",
            "description": "YYYY-MM-DD Colombia local",
            "schema": {
              "example": "2026-07-15",
              "type": "string"
            }
          },
          {
            "name": "zoneId",
            "required": false,
            "in": "query",
            "description": "Filtro por zona",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reportes"
        ]
      }
    }
  },
  "info": {
    "title": "FuturaGest API",
    "description": "",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}